Stop generating unique png files on form update
Instead of generating a unique png file when a form is updated,
overwrite the existing png file. This aligns with the way other file
formats are handled. Like other formats, the previous png files are
available in the tracking repo commit history and links to previous png
files remain valid because they are generated using commit sha absolute
link syntax.

Signed-off-by: Aaron Browne <aaron0browne@gmail.com>